backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
passlib[bcrypt]
python-jose
pydantic
